opt: add padding animation back handler
